web-ui-nuxt/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic~=2.9.2
pandas~=2.2.3
phonemizer~=3.3.0
supabase~=2.9.1
python-dotenv~=1.0.1